fix: Responsive FAQ Section
