+++ src/components/FaqSection/Faq.styles.jsx
@@ -1,15 +1,15 @@
 import styled from 'styled-components';
 import tw from 'twin.macro';
 
-
 const Heading = styled.div`
   ${tw`
         grid
         w-100
         grid-rows-1
         gap-4
+        pr-2
     `}
-`
+`;
 
 const FaqContainer = styled.div`
   ${tw`
@@ -17,52 +17,49 @@ const FaqContainer = styled.div`
         relative
         p-2
         rounded-lg
+        sm:mr-8
    `}
-`
+`;
 
 const Section = styled.div`
   ${tw`
-        grid
-        grid-flow-col 
-        grid-cols-2
-        w-full
-        gap-4
+        flex
+        lg:grid grid-cols-1 
+        md:w-100
+        gap-5
         mt-4
         text-white
     `}
-`
+`;
 
 const Bot = styled.div`
   ${tw`
-        m-4
-        flex
+        ml-4
+        mt-5   
     `}
-`
+`;
+
+const Img = styled.img`
+  ${tw`
+        float-left
+        mr-2
+        
+    `}
+`;
 
 const Chat = styled.div`
-  ${tw`
+  ${tw` 
         ml-3
         flex
-        items-center
-        justify-center
+        w-2/3
+        p-1.5
         bg-purple-light
         rounded-full
         relative
         text-black
         border-2 border-gray-800
     `}
-`
-
-const Bodydark = styled.p`
-  ${tw`
-        font-normal
-        text-base
-        text-black
-        font-roboto
-        py-1 
-        px-5
-    `}
-`
+`;
 
 const Ques = styled.div`
   ${tw`
@@ -71,7 +68,7 @@ const Ques = styled.div`
         mt-8
         mb-4
     `}
-`
+`;
 const Ul = styled.ul`
   ${tw`
         grid grid-flow-row auto-rows-max
@@ -83,10 +80,10 @@ const Ul = styled.ul`
         border border-gray-800
         mr-3 
     `}
-`
+`;
 
- const List = styled.li` 
- ${tw`
+const List = styled.li`
+  ${tw`
         p-4
         border-solid
         border-b border-gray-500
@@ -97,10 +94,10 @@ const Ul = styled.ul`
         font-roboto
         hover:text-white
     `}
-`
+`;
 
- const List1 = styled.li
-`${tw`
+const List1 = styled.li`
+  ${tw`
         p-4
         border-solid
         cursor-default
@@ -110,7 +107,6 @@ const Ul = styled.ul`
         font-roboto
         hover:text-white
     `}
-`
+`;
 
-
-export {List, List1, Ul, Ques, Bodydark, Chat, Bot, Section, FaqContainer, Heading};
+export { List, List1, Ul, Ques, Chat, Bot, Section, FaqContainer, Heading, Img };
