+++ src/components/FaqSection/FaqData.jsx
@@ -1,0 +1,18 @@
+const STAGE = {
+  QUESA: 'What is a Hackathon? How different is Online Hackathon?',
+  ANSA: 'ANSa ANSa ANSa ANSa ANSa ANSa ANSa ANSa ANSa ANSa ans ans ans ans ans ans ans ans ans an',
+  QUESB: 'How to register?',
+  ANSB: 'ANSb ANSb ANSb ANSb ANSb ANSb ANSb ANSb',
+  QUESC: ' Who can participate?',
+  ANSC: 'ANSc ANSc ANSc ANSc ANSc ANSc ANSc ANSc',
+  QUESD: 'What if I haven’t been to a Hackathon before?',
+  ANSD: 'ANSd ANSd ANSd ANSd ANSd ANSd ANSd ANSd',
+  QUESE: " What if I don't have a team or idea?",
+  ANSE: 'ANSe ANSe ANSe ANSe ANSe ANSe ANSe ANSe',
+  QUESF: 'Where can I find updates?',
+  ANSF: 'ANSf ANSf ANSf ANSf ANSf ANSf ANSf ANSf',
+  QUESG: '  What could be the size of a team?',
+  ANSG: 'ANSg ANSg ANSg ANSg ANSg ANSg ANSg ANSg',
+};
+
+export default STAGE;
