--- src/components/FaqSection/index.js
@@ -1,1 +1,0 @@
-export { default as Faq } from './Faq/Faq';
