+++ src/components/index.js
@@ -1,4 +1,4 @@
 // eslint-disable-next-line import/no-cycle
 export * from './marginals';
 export * from './shared';
-export * from './FaqSection'
+export { default as Faq } from './FaqSection/Faq';
