+++ src/components/shared/Typography/Heading5.jsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+import tw from 'twin.macro';
+
+export default styled.p`
+  ${tw`
+  font-normal
+        text-base
+        text-black
+        font-roboto
+        py-1 
+        px-5
+    `}
+`;
