+++ src/components/shared/index.js
@@ -4,5 +4,6 @@ export { default as Heading1 } from './Typography/Heading1';
 export { default as Heading2 } from './Typography/Heading2';
 export { default as Heading3 } from './Typography/Heading3';
 export { default as Heading4 } from './Typography/Heading4';
+export { default as Heading5 } from './Typography/Heading5';
 export { default as Body } from './Typography/Body';
 export { default as HighlightBody } from './Typography/HighlightBody';
